requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-js-asset==3.1.2
django-parler==2.3
djangorestframework==3.16.0
idna==3.10
parler==1.0.1
pillow==11.2.1
python-decouple==3.8
requests==2.32.3
rosetta==0.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
